fix(condo): DOMA-6786 used is32BitInteger in number filter
