+++ apps/condo/domains/common/utils/helpers.spec.ts
@@ -1,4 +1,11 @@
-import { getFiltersFromQuery } from '@condo/domains/common/utils/helpers'
+import { faker } from '@faker-js/faker'
+
+import {
+    getFiltersFromQuery,
+    is32BitInteger,
+    MAX_32BIT_INTEGER,
+    MIN_32BIT_INTEGER,
+} from '@condo/domains/common/utils/helpers'
 
 describe('Helpers property', () => {
     describe('getFiltersFromQuery', () => {
@@ -12,6 +19,32 @@ describe('Helpers property', () => {
             it('if invalid JSON is provided', () => {
                 expect(getFiltersFromQuery<{ [key: string]: string }>({ filters: '{"key": value, "key2": value}' })).toStrictEqual({})
             })
+        })
+    })
+
+    describe('is32BitInteger', () => {
+        it('value valid 32bit integer', () => {
+            const validInteger = faker.datatype.number({ min: MIN_32BIT_INTEGER, max: MAX_32BIT_INTEGER })
+
+            expect(is32BitInteger(validInteger)).toBeTruthy()
+        })
+
+        it('value less than min 32bit integer value', () => {
+            const notValidInteger = faker.datatype.number({ min: Number.MIN_SAFE_INTEGER, max: MIN_32BIT_INTEGER - 1 })
+
+            expect(is32BitInteger(notValidInteger)).toBeFalsy()
+        })
+
+        it('value greater than 32bit integer', () => {
+            const notValidInteger = faker.datatype.number({ min: MAX_32BIT_INTEGER + 1, max: Number.MAX_SAFE_INTEGER })
+
+            expect(is32BitInteger(notValidInteger)).toBeFalsy()
+        })
+
+        it('value not a number', () => {
+            const notInteger = faker.lorem.word()
+
+            expect(is32BitInteger(notInteger)).toBeFalsy()
         })
     })
 })
